OEL-713: Testing refactor.
